extracted calendar into own component
